solver.solver: move plotting functions to separate module
