+++ puncturedfem/solver/solver.py
@@ -6,9 +6,7 @@ Module containing the solver class, which is a convenience class for solving
 the global linear system.
 """
 
-import matplotlib.pyplot as plt
-from matplotlib.cm import ScalarMappable
-from numpy import inf, nanmax, nanmin, ndarray, shape, zeros
+from numpy import ndarray, shape, zeros
 from scipy.sparse import csr_matrix
 from scipy.sparse.linalg import spsolve
 from tqdm import tqdm
@@ -274,160 +272,7 @@ class solver:
             shape=(self.num_funs, 1),
         )
 
-    # PLOT SOLUTION ##########################################################
-
-    # TODO: this belongs in a different module
-
-    def plot_solution(
-        self,
-        title: str = "",
-        show_fig: bool = True,
-        save_fig: bool = False,
-        filename: str = "solution.pdf",
-        fill: bool = True,
-    ) -> None:
-        """
-        Plot the solution.
-        """
-        self.plot_linear_combo(
-            self.soln,
-            title=title,
-            show_fig=show_fig,
-            save_fig=save_fig,
-            filename=filename,
-            fill=fill,
-        )
-
-    def plot_linear_combo(
-        self,
-        u: ndarray,
-        title: str = "",
-        show_fig: bool = True,
-        save_fig: bool = False,
-        filename: str = "solution.pdf",
-        fill: bool = True,
-    ) -> None:
-        """
-        Plot a linear combination of the basis functions.
-        """
-        if not (show_fig or save_fig):
-            return
-        # compute linear combo on each cell, determine range of global values
-        vals_arr = []
-        vmin = inf
-        vmax = -inf
-        for cell_idx in self.V.T.cell_idx_list:
-            coef = self.get_coef_on_cell(cell_idx, u)
-            vals = self.compute_linear_combo_on_cell(cell_idx, coef)
-            vals_arr.append(vals)
-            vmin = min(vmin, nanmin(vals))
-            vmax = max(vmax, nanmax(vals))
-
-        # determine axes
-        min_x, max_x, min_y, max_y = self.get_axis_limits()
-        w, h = self.get_figure_size(min_x, max_x, min_y, max_y)
-
-        # get figure object
-        fig = plt.figure(figsize=(w, h))
-
-        # plot mesh edges
-        for e in self.V.T.edges:
-            plt.plot(e.x[0, :], e.x[1, :], "k")
-
-        # plot interior values on each cell
-        for cell_idx in self.V.T.cell_idx_list:
-            vals = vals_arr[cell_idx]
-            if vmax - vmin > 1e-6:
-                K = self.V.T.get_cell(cell_idx)
-                abs_cell_idx = self.V.T.get_abs_cell_idx(cell_idx)
-                K.parameterize(self.V.quad_dict)
-                if fill:
-                    plt.contourf(
-                        K.int_x1,
-                        K.int_x2,
-                        vals_arr[abs_cell_idx],
-                        vmin=vmin,
-                        vmax=vmax,
-                        levels=32,
-                    )
-                else:
-                    plt.contour(
-                        K.int_x1,
-                        K.int_x2,
-                        vals_arr[abs_cell_idx],
-                        vmin=vmin,
-                        vmax=vmax,
-                        levels=32,
-                        colors="b",
-                    )
-        if fill:
-            sm = ScalarMappable(norm=plt.Normalize(vmin=vmin, vmax=vmax))
-            plt.colorbar(
-                mappable=sm,
-                ax=fig.axes,
-                fraction=0.046,
-                pad=0.04,
-            )
-        plt.axis("equal")
-        plt.axis("off")
-        plt.gca().set_aspect("equal")
-        plt.subplots_adjust(
-            left=0.0, right=1.0, bottom=0.0, top=1.0, wspace=0.0, hspace=0.0
-        )
-
-        if len(title) > 0:
-            plt.title(title)
-
-        if save_fig:
-            plt.savefig(filename)
-
-        if show_fig:
-            plt.show()
-
-        plt.close(fig)
-
-    def get_axis_limits(self) -> tuple[float, float, float, float]:
-        """
-        Get the axis limits.
-        """
-        min_x = inf
-        max_x = -inf
-        min_y = inf
-        max_y = -inf
-        for e in self.V.T.edges:
-            min_x = self.update_min(min_x, e.x[0, :])
-            max_x = self.update_max(max_x, e.x[0, :])
-            min_y = self.update_min(min_y, e.x[1, :])
-            max_y = self.update_max(max_y, e.x[1, :])
-        return min_x, max_x, min_y, max_y
-
-    def update_min(self, current_min: float, candidates: ndarray) -> float:
-        """
-        Update the minimum value.
-        """
-        min_candidate = min(candidates)
-        new_min = min(current_min, min_candidate)
-        return new_min
-
-    def update_max(self, current_max: float, candidates: ndarray) -> float:
-        """
-        Update the maximum value.
-        """
-        max_candidate = max(candidates)
-        new_max = max(current_max, max_candidate)
-        return new_max
-
-    def get_figure_size(
-        self, min_x: float, max_x: float, min_y: float, max_y: float
-    ) -> tuple[float, float]:
-        """
-        Get the figure size.
-        """
-        dx = max_x - min_x
-        dy = max_y - min_y
-        h = 4.0
-        w = h * dx / dy
-        return w, h
+    # COMPUTE LINEAR COMBINATION #############################################
 
     def compute_linear_combo_on_cell(
         self, cell_idx: int, coef: ndarray
